// legacy/schema.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      products: {
        Row: {
          attributes: { [key: string]: string }[] | null
          brand: string
          category: string
          created_at: string
          description: string
          featured: boolean | null
          frame_style: string | null
          gender: string | null
          has_variants: boolean
          id: string
          images: string[]
          instock: boolean | null
          price: number
          slug: string
          sub_category: string
          title: string
          variants: { variant_name: string; variant_value: string }[] | null
        }
        Insert: {
          attributes?: { [key: string]: string }[] | null
          brand: string
          category: string
          created_at?: string
          description: string
          featured?: boolean | null
          frame_style?: string | null
          gender?: string | null
          has_variants?: boolean
          id?: string
          images: string[] | null
          instock?: boolean | null
          price?: number
          slug: string
          sub_category: string
          title: string
          variants?: Json | null
        }
        Update: {
          attributes?: { [key: string]: string }[] | null
          brand?: string
          category?: string
          created_at?: string
          description?: string
          featured?: boolean | null
          frame_style?: string | null
          gender?: string | null
          has_variants?: boolean
          id?: string
          images?: string[] | null
          instock?: boolean | null
          price?: number
          slug?: string
          sub_category?: string
          title?: string
          variants?: Json | null
        }
        Relationships: [
          {
            foreignKeyName: "products_brand_fkey"
            columns: ["brand"]
            isOneToOne: false
            referencedRelation: "brands"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "products_category_fkey"
            columns: ["category"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "products_sub_category_fkey"
            columns: ["sub_category"]
            isOneToOne: false
            referencedRelation: "sub_categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
// ...
}